Adapt all instances of the term frontend to the new term (language) module.
