+++ languages/kotlin/README.md
@@ -1,6 +1,6 @@
-# JPlag Kotlin language frontend
+# JPlag Kotlin language module
 
-The JPlag Kotlin frontend allows the use of JPlag with submissions in Kotlin. <br>
+The JPlag Kotlin module allows the use of JPlag with submissions in Kotlin. <br>
 It is based on the [Kotlin ANTLR4 grammar](https://github.com/antlr/grammars-v4/tree/master/kotlin/kotlin), licensed under the Apache 2.0.
 
 ### Kotlin specification compatibility
@@ -14,13 +14,13 @@ If there are any major updates or fixes to the grammar<a href="#footnote-1"><sup
 
 ### Token Extraction
 
-The choice of tokens is intended to be similar to the Java or C# frontends. It includes a range of nesting structures (class, method, control flow expressions) as well as variable declaration, object creation, assignment, and control flow altering keywords.
+The choice of tokens is intended to be similar to the Java or C# modules. It includes a range of nesting structures (class, method, control flow expressions) as well as variable declaration, object creation, assignment, and control flow altering keywords.
 
 More syntactic elements of Kotlin may turn out to be helpful to include in the future.
 
 ### Usage
 
-To use the Kotlin frontend, add the `-l kotlin` flag in the CLI, or use a `JPlagOption` object set to `LanguageOption.KOTLIN` in the Java API as described in the usage information in the [readme of the main project](https://github.com/jplag/JPlag#usage) and [in the wiki](https://github.com/jplag/JPlag/wiki/1.-How-to-Use-JPlag).
+To use the Kotlin module, add the `-l kotlin` flag in the CLI, or use a `JPlagOption` object set to `LanguageOption.KOTLIN` in the Java API as described in the usage information in the [readme of the main project](https://github.com/jplag/JPlag#usage) and [in the wiki](https://github.com/jplag/JPlag/wiki/1.-How-to-Use-JPlag).
 
 <br>
